Add files via upload
This project is a spam classifier that uses NLTK (Natural Language Toolkit) and Naive Bayes algorithm to classify emails and SMS as spam or not. The dataset used for training and testing the model was preprocessed by performing data cleaning and visualizations. The model was deployed using Streamlit library to convert it into a website form.

The code for this project is available on GitHub, and it can be used to classify emails and SMS as spam or not. The project is useful for people who want to automate the process of identifying spam emails and SMS. Additionally, the code can be modified and improved to make it more accurate and efficient.

Overall, this project demonstrates how to use NLTK, Naive Bayes algorithm, data cleaning, visualizations, and Streamlit library to create a spam classifier.